Use single file than multiple files
